backend/pyproject.toml
[project]
name = "neoforge"
version = "0.1.0"
description = "NeoForge Starter Kit"
authors = [
    {name = "NeoForge Team", email = "team@neoforge.dev"},
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "prometheus_client>=0.19.0",
    "email-validator>=2.1.0",
    "fastapi-mail>=1.4.1",
    "PyJWT>=2.8.0",
    "psutil>=5.9.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-api>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.47b0",
    "opentelemetry-instrumentation-redis>=0.47b0",
    "opentelemetry-instrumentation-httpx>=0.47b0",
    "aiohttp>=3.9.3",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.4",
    "pytest-cov>=4.1.0",
    "pytest-watch>=4.2.0",
    "factory-boy>=3.3.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"
